Organize folders
Organize folders by purpose to make it easier to read.
