+++ src/containers/PlanetsContainer.test.jsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import PlanetsContainer from './PlanetsContainer';
 
-jest.mock('./assets');
+jest.mock('../assets');
 
 describe('PlanetsContainer', () => {
   const dispatch = jest.fn();
